refactor: merge same function action
